Do not duplicate scalac options in partest
This fixes the incorrect result for the run/macro-settings test
mentioned in the previous partest rewrite.

The CompilerCommand constructor adds "opts" to "testSettings", and so did
the check which resulted in two "-Xmacro-settings:hello=1" options passed
to scalac which in turn caused the unexpected output:
List(hello=1, hello=1)

The argument processing result is already collected by CompilerCommand,
so it can be reused here.